structure: move server/art/ to server/services/art/
Art fetching and caching is domain logic over the DB like the other
services; the HTTP route stays at server/routes/art/.
