+++ server/services/art/epicIcon.test.ts
