+++ server/services/art/epicIcon.ts
