+++ server/services/art/fetch.ts
