+++ server/services/art/invalidate.ts
