+++ server/services/art/missing.ts
