+++ server/services/art/paths.ts
